backend-tests: add email verification to Test2FAEnterprise
With recent changes in the useradm-enterprise it is not possible
to enable/disable two factor authenitcation without email address verification.
Refactor the tests to use common smtp_mock.
Wait for the port release in the smtp mock.

changelog: none
